requirements.txt
aiohttp==3.13.3
aiosqlite==0.22.1
discord.py==2.6.4
Pillow==12.1.1
psutil==7.2.2
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
